glyphsync_comfyui_pip_package/requirements.txt
numpy
Pillow
easyocr
